Sync config/service on FreeBSD
The FreeBSD openvpn init script assumes that the service name is the
same as the config name.  By using a different name for the service
(`openvpn_foo`) and the configuration file (`foo.conf`), we must provide
an additional variable in `rc.conf.d/openvpn_foo`, `rc.conf.local` or
`rc.conf` (the first one start exists).

~~~
openvpn_foo_configfile=/usr/local/etc/openvpn/foo.conf
~~~

This file is managed by the service resource from Puppet that does not
offer an opportunity to define such a variable at the right place in a
convenient way, so switch back to not prefixing the service name with
`openvpn_`.